Remove low latency feature from waterfall field trial config
We want to launch GPU vsync and presentation feedback on Windows on all
channels and do a separate experiment for low latency presentation.
Therefore this CL removes the low latency presentation finch flag from
the waterfall config so that it matches the server finch config which
will also remove any experiment associated with that flag, and we don't
get a presubmit warning for the server config.  Also in the future when
we experiment with this flag separately we don't want conflicts with the
existing config either on the waterfall or on the server.

Bug: 953970
Change-Id: I38d792298b41bc7877d09bb454bd28ab3894fb82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954564
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Steven Holte <holte@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#722727}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8091bcdfc8775c837356bf1f65a5b9ef57f93e31