Added multiplier error margin
Instead of a fixed 2ms margin of error for frame_times. Now using a
percentage (10%) of upper limit.
So for a story with upper limit of 16ms margin will be 1.6ms
and for a story with upper limit of 60 ms would be 6ms.

TBR=crouleau@chromium.org,sadrul@chromium.org

Bug: chromium:1031609
Change-Id: I67ceb4f84af5d07ce5606856afd868265060e6a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956042
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#722750}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5eb683cf7fa8671dc9e64a73a48047d7180f1399