[Perf Waterfall] Always grab webview shell from next to primary apk.
Telemetry already has logic to grab the shell apk that is next to the
primary apk (whether it be monochrome or system webview). There is no
reason to provide an additional full-length path to the shell apk.
Adding this flag is complicated because it must be separately added
for pinpoint and for the perf waterfall, or otherwise they start
with different arguments. Removing it simplifies things and ensures
that we don't accidentally run a shell apk that is from a different
chromium output directory from the primary apk.

See this discussion for details:
https://groups.google.com/a/google.com/forum/#!topic/chrome-benchmarking/LMAJAgjftZo

Also, no longer allow --browser=exact for webview. per
https://chromium-review.googlesource.com/c/chromium/src/+/1959661/3/testing/buildbot/chromium.perf.json#1337
if you pass --browser=exact and --browser-executable=.../monochrome.apk
then Telemetry will run Clank, not webview. This has always been true
and is true regardless of whether the --webview-embedder-apk flag is passed.

Bug: 1028125
Change-Id: I9ac19cc7b2fab586ccf19c0b5eb7838152486326
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1962908
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#723974}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c1324fa4ee0802370abbd624df217ef7b479180c