[Benchmarking] Sharded gtests ignores passthrough arguments.
The arguments that the run_performance_tests.py script does not recognize are
by default passed through to the Telemetry executable. Now that we have gtest
on the same shard, those passthrough arguments are confusing the gtest perf
tests.

Example:
https://chrome-isolated.appspot.com/browse?namespace=default-gzip&digest=0e33146297150a5de344790520644a01710927c7&as=benchmark_log.txt
is broken because of the added arguments
 -v --browser=android-chromium --upload-results --assert-gpu-compositing
. For Gtest you can instead add flags to the shard itself.
In the future, we should figure out how to put the passthrough arguments
onto the shard map for Telemetry as well, but for now this will fix the
immediate problem.

Bug: 921353
Change-Id: Ie5d853aa402fc84e21a4f2f3cf180cb150f7e428
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1981100
Commit-Queue: Caleb Rouleau <crouleau@chromium.org>
Auto-Submit: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#728088}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 821af7a52ab766aebf2ae5edb591e85c50c0c0f7