Add AssumeOverlapAfterFixedOrStickyPosition trial
Based on early results in dev/canary, this feature is looking promising
and we would like to continue this experiment in Beta.  To unblock that
we are enabling this experiment for the perf bots now to get more data
on the impact.

Bug: 752756
Change-Id: Ia8078787194ce2ee5e1922456dc80576f84ef176
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993144
Commit-Queue: Sam Fortiner <samfort@microsoft.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731698}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fec62b63db24b5ed7c4eaa1ab3a0808e497112df