Makes all browsertests set kUseFakeDeviceForMediaStream
Content's TestLauncherDelegate was doing this, but that isn't
called in all paths and there was a lot of tests setting the
switch as well.

Some tests need to change the value, which are updated to remove
the switch and then set it.

BUG=none
TEST=none

Change-Id: Ia7be97b01eeadd35fba9b31906f120ef58b8e99e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007524
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#734027}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 879eaa73a0de0de776c564ece3b1e4286abdc447