If control test is noisy, failures are not reported
For the test to not fail when the control test is noisy, all failures
are marked as SKIP and expected. This will help with reducing the false
positives because we only report the failures which we are confident
about.

Bug: chromium:1029846
Change-Id: Icc7814813c58d94749fdbcf426df4eeaa5b2556b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2042639
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#739992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9533137034161a9f615139e3caa7aaacacafcadb