Disable PrivacySettingsRedesign experiment in fieldtrial_testing_config
This experiment was enabled in fieldtrial_testing_config, because it
was being rolled out in Chrome M80.

However, the experiment with a same name is being reused in M81/M82
code, where it's disabled again for the time being, and so it should be
disabled in fieldtrial_testing_config too.

Bug: 1032584
Change-Id: Ic903f1c082350d7067aaadbafee49b73772a21a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2064998
Commit-Queue: Martin Šrámek <msramek@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743175}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84948d6f9fba5a6a3512d9998e46e2a20aa8f658