Use Confidence interval only for Control test
In a previous change(crrev.com/2055047) we stop using the confidence intervals as a result
of high noise on them. But we need to use the CI for control test. This
change adds the condition to use confidence interval only for the
control test. This helps with not flagging flaky failures.

TBR=crouleau@chromium.org

Bug: chromium:1052054
Change-Id: Ibbacb64229d06fc8b896cce28eb07d41bed62885
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2065112
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Reviewed-by: Behdad Bakhshinategh <behdadb@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#743238}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: feffdfcaaac2e3b5db9db176ecf25097600a9cc1