Add CompositeCrossOriginIframes trial
Based on early results in dev/canary, this feature is looking promising
and we would like to continue this experiment in Beta.  To unblock that
we are enabling this experiment for the perf bots now to get more data
on the impact.

Bug: 1014273
Change-Id: I0b3c0dbb94075bdb94b3f770bd32f2e761c8f6c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2089953
Reviewed-by: Stefan Zager <szager@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Auto-Submit: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#747365}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d49f0f21eec162aee80c5999a18540cfb205178e