Use a whitelist for Mac OS static initializer checks.
When a dsym is present we can get the file that each initializer was from
and compare against a list of accepted files. This helps support code
coverage on the CQ, which adds 2 additional static initializers.

When a dsym is not available, we fall back to counting how many static
initializers there are and comparing that against a hardcoded number.

Bug: 1059432

Change-Id: I6a1df68371a34724822343198afb14fdd1f9cff8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2106349
Commit-Queue: Sajjad Mirza <sajjadm@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751397}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26561f3f44e985fcd316cbfa3f171dadde739f56