weblayer: removes support for --run-web-tests
This was used to trigger different modes in the renderer. This
wasn't the right way to do this, rather tests should use a
special fieldtrial. This wasn't working because finch wasn't wired
up. Now that finch is wired up the right config is being loaded and
this code is not necessary (and wrong).

BUG=1049426
TEST=covered by wpt

Change-Id: Ifbe095a29033511b506395b195cc12e0485f8122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109146
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751680}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 054707c2573902d87df8ff4e2833ccf0b60eca4b