weblayer: remove timeout-multiplier
A timeout multiplier of .25 means tests are expected to complete in
1/4 the time. This causes timeouts in some tests. We shouldn't
need to explicitly set a multiplier, and if there is a need to it
should be done in the wpt runner.

This also removes HOST_RESOLVER_ARGS as it is unused.

BUG=1049426
TEST=test only change

Change-Id: Ifbefc33cd9c595d689f4fb7500e61459d20c7c6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2109022
Reviewed-by: Andrew Luo <aluo@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Auto-Submit: Scott Violet <sky@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#751739}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c17ff9872a5c0c642665df8c9d68f97d060ecf8d