Run FYI Mac non-isolated scripts in compile task
These non-isolated scripts are fast and only need to be run once,
not on every OS version that we run Mac tests on. This change is
a test to see if it is reasonable to run these scripts directly on
the task which produces the binaries, rather than running them on
the OS-version-specific triggered testers.

This comes with a few advantages:
* No need to zip up all the build artifacts and transfer them to
  a tester task; the tester will only babysit swarming isolated
  tests, so there's no need for an archive build step.
* Testers can be non-Mac, since all they're doing is babysitting
  isolated tasks, allowing us to use cheap Linux GCE VMs for the
  testers.
* Unity of configs. We won't have to ensure that just one of the
  various testers has these scripts enabled; they'll just be
  enabled on the builder and nowhere else.

If this change works well, a follow-up CL will do the same for
Mac Builder on the main waterfall.

Bug: 1057776
Change-Id: I0f4150499ed09fc958c63b4bb9f3f7853f82f746
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159542
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Aaron Gable <agable@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#761238}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c8025db6ccc61ff0934cc0a1a98c71ffbe89edfe