Move Storage Service tests to CQ; sandboxed to FYI
Unsandboxed OOP Storage browser (etc) tests have been on FYI bots
for a while with no interesting issues.

Currently the relevant feature flag is in Finch for non-stable users,
and getting this on the CQ is part of the progression toward the feature
being on by default, at which point these steps can be removed.

In addition, the unsandboxed tests on FYI bots are switched over to
have sandboxing enabled, in preparation for the next phase of
development and rollout.

Bug: 994911
Change-Id: Ic8ab21116cc535e4656be782408f438ef043225a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2159569
Reviewed-by: Aaron Gable <agable@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#765117}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b63ede8826128c8a4f1a8b160e72c8cc6a84a14c