Focus textfields on completed tap on ChromeOS
A flag was introduced earlier to change the focus behavior of Views
textfields. Previously, textfields took focus on tap down. With the flag
enabled, they take focus when a tap is completed (i.e. when the touch is
lifted, and the gesture was interpreted as a tap).

This flag is being enabled through two separate Finch experiments: one
targeting ChromeOS (along with the WebUITabStrip for tablet devices),
and one targeting other desktop platforms. This updates the
fieldtrial_testing_config.json to also enable it on the waterfall and
updates tests accordingly.

Bug: 1069634
Change-Id: I6dd7693fd28e6c65c8a6a1f62657c28e7b4beca1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2170561
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#768427}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d9f81971b0b7ad93b856787e653cc0b1ffe7fd91