Add MaxOverlapBoundsForFixed field trial
Based on early results in dev/canary, this feature is looking promising
and we would like to continue this experiment in Beta.  To unblock that
we are enabling this experiment for the perf bots now to get more data
on the impact.

Bug: 1058987
Change-Id: I5e771beb6d4085c25df353c68c37db05d220af65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2211019
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Sam Fortiner <samfort@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#771357}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a7274d4f41f27675b67991b839fca7584f1991c