Add field trial config for AndroidWideColorGamut
Android uses the dynamic color gamut to change gamuts as needed based
on the content being displayed. There is a Finch experiment for this
which is rolled out to 50% of Canary and Dev channels.

This CL adds a field trial for that Finch experiment so bots can test
the config.

Change-Id: I112a9dc0f8edb99b39f227a81a7785527717db95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225622
Commit-Queue: Chris Blume <cblume@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#774393}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 75459e66dda89471434ac3bc567c242fb0489fd8