xvfb: Wait for ReparentNotify before executing tests.
X11 tests can fail because start up of OpenBox is not
synchronous and it is unknown when exactly it starts up.

That results in a client that uses X11 missing events
and flakiness of tests.

Thus, to avoid that issue, use a small program that
creates a dummy X11 windows and waits for the
ReparentNotify event. That is, OpenBox decorates
all the windows once it is initialized that results
in that event. Once the event is received, we can
be sure that the WM is up and running, and tests
can be executed.

Bug: 1078771
Change-Id: Ic6fae30a706a6b9c32e4670da56457c833b4b7f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214526
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#775001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2fb9ab545d61adb0b35a448957ae7645140f8220