[pgo] skip validation flag for profraw files
IR PGO isn't compatible with the way the workflow converts .profraw
files to a .profdata to validate the profraw results. A skip
validation flag is introduced to bypass the entire conversion step.
A warning is logged to ensure the user is aware of this. All found
profraw files are directly passed to llvm-profdata to merge into
the test.profdata file for that benchmark.

Bug: 1091310
Change-Id: Id67b7fd4e03d36baadf0b2d728cf5f8f14bc9682
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2231451
Commit-Queue: Jeff Yoon <jeffyoon@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#775361}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4d86a5e7ddfb90dd163334350963c0c573a4412a