Reporting the results only for initial run
For reducing the time of representative perf test, we only rerun the failed stories
and update the results of the initial run based on the reruns. If the
failed stories pass in rerun, the output will be updated to report all
tests as passing.

The current bug in these tests is that it would report values before
considering the rerun results. So even if the failed tests pass after
rerun, the output is still reporting the failures. This would cause the
return_code to be set as 1 (Failed/Interrupted) while the tests has
already passed via rerun.

Bug: chromium:1091743
Change-Id: I9eef7250bc2dad649fef49cf9a5ccd8195190f80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2232603
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#775682}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0063a381fa4f27cf64e2028fb76646b6cbe942bc