[WebLayer] Run weblayer_unittests on Android
The weblayer unittests needed a small fix to be able to run on Android,
these should be running on the bots. This will be useful for adding site
isolation unittests, which should be tested on both Android and desktop.

Bug: 1058364
Change-Id: I7188a6140cff2811594bc7ab6d2e9b2dd58366ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2235941
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#776359}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: be6657ec4f69e22b4a782a5c588621ed9e185cd4