X11: refactor DWTHX11 tests and move them to right places
This also adds a xwmstartupcheck program that resolves
the race between bringing up openbox and starting tests.

That is, we tried to enable the program for all the tests,
but after having a couple of tries, it turned out that
it's not trivial to figure out correct bots to use that.
In the end, the xvfb script fails because it is unable
to find that small program. Thus, it was decided that
only the x11_unittests target will depend on that program,
and once dpranke@ reworks the test() template, we will
rework the solution so that all the tests that use
x11 will use that small program to ensure there is not
race between the openbox and the tests.

Bug: 990756
Change-Id: I5b1a8504e54fc4f3a2172978e9d4604145d56847
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2228652
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#778692}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4d1cfbebf33f9b0943281a24b0ffbbafa680c15