[TTS] Add a balloon promo for Longpress.
Adds an IPH for switching from Tap triggering to Long-press triggering
as part of the ContextualSearchLongpressResolve experiment.

There are two strings that can show in the bubble, one for users that
have opened the panel in the past, and one for the general user-base
that has not opened the panel recently.

The IPH triggering logic is:
1) Show when the user taps and we would have tap-triggered.
2) Only show 3 times in a 90 day window.
3) Don't show if the user has done any longpress selection (which
   triggers Contextual Search) in the 90 day window.
4) Don't show if the user has opened the panel after a longpress in a
   1 year window (the max allowed).

BUG=1086724

Change-Id: I19335076afeb8a349c77e93a794d9adf8bfcfb95
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2233861
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#780030}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef6aca86713fcd919677178476d751c7f276bb5f