Oobe: Read Rollback Restore from Config
In order to restore the enrollment state after rollback, the oobe
flow needs to execute the enrollment screen.

Currently, this is not happening because the fact that a rollback
took place is not read from the oobe configuration.

Reading the correct flag is added in two places: To recover the
enrollment and to restart chrome afterwards in order to get the device
policies.

Bug: 1093928
Change-Id: Ib463dae4b1ac29e7f2ab870ae3b839f2018e11e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273159
Reviewed-by: Denis Kuznetsov [CET] <antrim@chromium.org>
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#783962}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91ef03171386058e4c821ab4ff544c168898e55f