[PwdCheckAndroid] Configure trybots to execute controller tests
Purpose of the testing/buildbot/* changes is the automated execution of
the newly introduced robolectric test target password_check_junit_tests
on trybots.

It was split out from https://crrev.com/c/2275959 to separate the bot
configs from the scaffolding that other CLs depend on. To repeat the
answers on previously asked questions:

> Has this test suite been run anywhere?

On other bots? No. SO far, it consists of one test that I only ran
locally.


> Do you know how long it takes?

The runtime of the entire suite will always be measures in seconds, not
minutes. The robolectric tests require no full browser instance and no
emulator which causes most tests to run subsecond.


> Is it stable?

Yes. It mocks everything that isn't part of the feature and uses neither
UI nor native code which increases stability over regular JUnit tests.


> What are the plans for the test suite long term? Will it end up with a
lot of logic in it?

The scope of this target is limited to exactly the password check
feature. That feature is comparable in scope to
touch_to_fill_junit_tests or keyboard_accessory_junit_tests. Compared to
these components, the password check is even smaller.

Bug: 1101256,1092444
Change-Id: I8398abec3a10502f23dc0140f7080b694e26baa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283318
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788155}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35d4eda4cf305cc7cf7016cf117539f1fa1e4f90