Decrease primary task slice expiration from 10m to 5m for CrOS HW tests.
Currently, CrOS HW tests will wait 10min for a device on an up-to-date
OS, and fallback to a device running any OS after that.

With crbug.com/1057152, we are now flashing the device to the right OS
as part of the task. So we no longer need to wait so long for the right
device when we can flash it ourselves. Flashing takes about 5min, so a
5min deadline maintains a 10min delay at the worst.

This also moves the expiration value from an arg to a module constant
since the arg was never passed, and likely won't ever be.

Bug: 1057152
Change-Id: Ia66e714b02f9f4b6b0d011fd96a80f42d0294d52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287954
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790145}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fdd4a770b2cc72676fee29ddd4db5cfb475501aa