Re-add field trial testing config for DisableLatencyRecoveryDesktop
Based on the data gathered from running the experiment on canary/dev and
beta, it seems that the disabling latency recovery for both main and
impl threads is an overall performance win. So, we are going to
re-enable the (now-expired) experiment on stable and move forward with
its gradual rollout.

Bug: 993895
Change-Id: I5c45c3cb469d14cf53193d0af6a560842ce02be2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2309188
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790549}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26f4c62b13250ba0f39157573c78ad10adc1821a