Add --use-gpu-in-tests flag to webgpu blink web tests.
Previously tests were passing because
we would manually discover adapters all adapters on the system
and then pick the first one, which just so happens to be the
one compatible with ANGLE (when running Chrome normally).

We are changing this so that we explicitly
pick the adapter that ANGLE is using.
Since ANGLE isn't initialized without --use-gpu-in-tests,
the WebGPU tests would all fail with no available adapter.

We will work on hooking up SwiftShader in the future.

Change-Id: Ifecab52bed926d0abee9bb161565e77aeee52885
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2310391
Commit-Queue: Natasha Lee <natlee@microsoft.com>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Austin Eng <enga@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#790677}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fad33e98dfff9dbb57d1df0c7dd3c9e3aaf09194