[lacros] LUCI config for linux-ash-chromium-builder-fyi-rel
This linux-ash-chromium-builder-fyi-rel is created to upload ash-chrome
continuously to gs and the binary will be used to test lacros-chrome.
The upload logic will be in the next cl when the bot can launch and run.

Bug: 1109287
Ignore-CQ-Freeze: True
Change-Id: Ib8c2983a84d8a16510634c29df2ee1680438651f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327372
Commit-Queue: Sven Zheng <svenzheng@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#792936}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 50d3d14e3cbe04cc24067aec02b5a9a3d6b0e9db