chromeos: Don't swallow test args when '--' is present.
When '--' is present on the cmd line, anything to the right of it is
considered args to the test binary and will be pass untouched to it.

This also adds the '--' arg to all isolated-script tests, which is
practically only telemetry tests.

Bug: 1108412
Change-Id: Iec53e68d39c0bbe947ebdc166426d6b4c848e892
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2318309
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#794225}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 983340cc7c5746b3619b4d4bddbfa2340c4219af