[lacros] LUCI config for linux-lacros-builder-rel and linux-lacros-tester-rel
Adding lacros builder/tester to CI with not closing tree option.
The selected tests are tests running in the fyi bot and are stable:
https://ci.chromium.org/p/chromium/builders/ci/linux-lacros-tester-fyi-rel
When the bot is stable, we will change it to closing tree to true.

Bug: 1104291
Change-Id: I8ff14f3cba6b0bcca256a100beb6d7a80362672e
Ignore-CQ-Freeze: True
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333334
Reviewed-by: Erik Chen <erikchen@chromium.org>
Reviewed-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Sven Zheng <svenzheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#794277}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 577fe5ff7c05b8dc4c7bc2597c6ab57a6c3f91b8